Rename validator_custody.rs ->custody_context.rs
