+++ beacon_node/beacon_chain/src/custody_context.rs
